requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
asyncio==3.4.3
asyncpg==0.30.0
attrs==24.3.0
Backtesting==0.3.3
beautifulsoup4==4.12.3
blinker==1.9.0
bokeh==3.6.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cli_helpers==2.3.1
click==8.1.7
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
fastapi==0.115.6
finta==1.3
fonttools==4.55.7
frozendict==2.4.6
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
html5lib==1.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
multitasking==0.0.11
narwhals==1.20.1
numpy==2.2.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pandas-ta==0.3.14b0
peewee==3.17.8
pgcli==4.1.0
pgspecial==2.1.3
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.29.2
psycopg==3.2.3
psycopg-binary==3.2.3
pulsar-client==3.5.0
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
pyotp==2.9.0
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.15.0
selenium==4.27.1
setproctitle==1.3.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
starlette==0.41.3
streamlit==1.41.1
ta-lib @ file:///C:/ta-lib/ta_lib-0.5.2-cp311-cp311-win_amd64.whl
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
upstox-python-sdk==2.12.0
urllib3==2.2.3
uuid==1.30
uvicorn==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
wsproto==1.2.0
xyzservices==2024.9.0
yfinance==0.2.52


httpx==0.25.0
pydantic-settings==2.0.3
cryptography

